Extract configuration file from build
