+++ examples/blog/index.html
@@ -5,11 +5,12 @@
     <title>React-Admin Blog Sample Application</title>
 </head>
 <body>
-    <script src="bower_components/fakerest/dist/FakeRest.min.js" type="text/javascript"></script>
-    <script src="bower_components/sinon-server/index.js" type="text/javascript"></script>
+    <script src="/bower_components/fakerest/dist/FakeRest.min.js" type="text/javascript"></script>
+    <script src="/bower_components/sinon-server/index.js" type="text/javascript"></script>
     <script src="data.js" type="text/javascript"></script>
     <script src="fakerest-init.js" type="text/javascript"></script>
 
     <script src="http://localhost:8080/bundle.js"></script>
+    <script src="config.js"></script>
 </body>
 </html>
